pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "warp-content-processor"
version = "0.1.0"
description = "Content processor for Warp Terminal"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Warp Content Processor Contributors" },
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0",
    "pylint>=2.17.5",
    "ruff>=0.1.9",
    "mypy>=1.5.1",
    "types-PyYAML>=6.0.12",
    "bandit>=1.7.5",
    "yamllint>=1.32.0",
    "pip-audit>=2.7.0"
]

[project.urls]
homepage = "https://github.com/tzervas/warp-content-processor"
repository = "https://github.com/tzervas/warp-content-processor.git"

[tool.hatch.build.targets.wheel]
packages = ["src/warp_content_processor"]

[tool.hatch.build.targets.sdist]
packages = ["src/warp_content_processor"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
]

[tool.pylint.format]
max-line-length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
